Refactor app
 - Keep run dependencies in a server
 - Move handler to individual file
 - Add method to get run address
