+++ searcher.go
@@ -1,0 +1,31 @@
+package main
+
+import (
+	"fmt"
+	"index/suffixarray"
+	"io/ioutil"
+)
+
+type Searcher struct {
+	CompleteWorks string
+	SuffixArray   *suffixarray.Index
+}
+
+func (s *Searcher) Load(filename string) error {
+	dat, err := ioutil.ReadFile(filename)
+	if err != nil {
+		return fmt.Errorf("Load: %w", err)
+	}
+	s.CompleteWorks = string(dat)
+	s.SuffixArray = suffixarray.New(dat)
+	return nil
+}
+
+func (s *Searcher) Search(query string) []string {
+	idxs := s.SuffixArray.Lookup([]byte(query), -1)
+	results := []string{}
+	for _, idx := range idxs {
+		results = append(results, s.CompleteWorks[idx-250:idx+250])
+	}
+	return results
+}
